requirements.txt
# ==========================
# 🚀 Streamlit LangChain + LlamaIndex Chatbot
# Production-Stable Requirements (Tested Oct 2025)
# ==========================

# --- Core Framework ---
streamlit==1.39.0
python-dotenv==1.0.1

# --- LangChain Ecosystem (Stable) ---
langchain==0.3.7
langchain-openai==0.2.6
openai==1.51.2

# --- LlamaIndex Ecosystem (Latest Stable) ---
llama-index==0.11.20
llama-index-core==0.11.20
llama-index-llms-openai==0.2.18
llama-index-embeddings-openai==0.2.1

# --- Embeddings, Tokenization & Utilities ---
tiktoken==0.8.0
numpy==1.26.4
packaging==24.1

# --- Optional (Logging, Compatibility) ---
tenacity==8.3.0
requests==2.32.3
